Route register
Grouping and routing of all processes associated with a user's registration in a file different from the server.js file
